feat : 바우처 사용 기능 구현 (#61)

+++ ourhome-market/src/main/java/com/armand/ourhome/market/voucher/domain/FixedVoucher.java
@@ -40,7 +40,7 @@ public class FixedVoucher extends Voucher {
   }
 
   @Override
-  public int getDiscountPrice(int currentPrice) {
+  public long getDiscountPrice(long currentPrice) {
     return this.amount;
   }
